v0.7.4: refactor generateCampaignPitch into modular pipeline (core, sections, assembly, cleanup, safety) with no intentional output changes
